+++ shaders/discretization.glsl
@@ -2,10 +2,10 @@
 
 layout (local_size_x = 8, local_size_y = 4, local_size_z = 1) in;
 layout (binding = EFTDEM_HEIGHTMAP_BUFFER) restrict buffer mapBuffer{
-    double heights[];
+    float heights[];
 };
 layout (binding = EFTDEM_CLOSING_MASK_BUFFER) restrict buffer discreteValuesBuffer{
-    double discreteValues[];
+    float discreteValues[];
 };
 
 uniform uvec2 resolution;
